test: add unit test suite and configure GitHub Actions CI workflow
- Format codebase to pass forge fmt --check CI step
- Refactor test/Bank.t.sol: add receive() fallback to handle ETH transfers,
  add test cases for admin withdrawal and access control
- Add test/BigBank.t.sol: test deposit modifier, receive() fallback, changeAdmin,
  two-step admin withdrawal pipeline, Top 3 ranking, event assertions, fuzz tests

+++ src/BigBank.sol
@@ -1,55 +1,55 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.13;
 
-import {Bank,IBank} from "./Bank.sol";
+import {Bank, IBank} from "./Bank.sol";
 
-contract BigBank is Bank{
+contract BigBank is Bank {
     address public immutable owner;
 
-    constructor(){
-        owner=msg.sender;
+    constructor() {
+        owner = msg.sender;
     }
 
-    modifier depositAmountGreaterThan001(){
-        require(msg.value>0.001 ether, "Deposit amount must greater than 0.001 ether");
+    modifier depositAmountGreaterThan001() {
+        require(msg.value > 0.001 ether, "Deposit amount must greater than 0.001 ether");
         _;
     }
 
-    function deposit() external payable override depositAmountGreaterThan001{
+    function deposit() external payable override depositAmountGreaterThan001 {
         _handleDeposit();
     }
 
-    receive() external payable override{
+    receive() external payable override {
         require(msg.value > 0.001 ether, "Deposit amount must greater than 0.001 ether");
         _handleDeposit();
     }
 
     function changeAdmin(address newAdmin) public {
         require(msg.sender == owner, "Only owner can change admin");
-        require(newAdmin != address(0),"New admin can not be zero address");
+        require(newAdmin != address(0), "New admin can not be zero address");
         admin = newAdmin;
     }
 }
 
-contract Admin{
+contract Admin {
     address public immutable admin;
 
     receive() external payable {}
 
-    constructor(){
+    constructor() {
         admin = msg.sender;
     }
 
-    function adminWithdraw(IBank ibank) external{
+    function adminWithdraw(IBank ibank) external {
         require(msg.sender == admin, "Only admin can withdraw");
         ibank.withdraw();
     }
 
-    function withdrawToOwner() external{
+    function withdrawToOwner() external {
         require(msg.sender == admin, "Only admin can withdraw");
-        uint balance = address(this).balance;
-        require(balance > 0,"No balance can to withDraw");
-        (bool success, ) = admin.call{value: balance}("");
+        uint256 balance = address(this).balance;
+        require(balance > 0, "No balance can to withDraw");
+        (bool success,) = admin.call{value: balance}("");
         require(success, "Withdraw failed");
     }
 }
